add quantity updated

+++ models/product_product.py
@@ -3,4 +3,4 @@ from odoo import models, fields
 class ProductProduct(models.Model):
     _inherit = "product.product"
 
-    minimum_qty = fields.Float(string="Minimum Order Quantity", default=1.0)
+    minimum_qty = fields.Integer(string="Minimum Order Quantity", default=1)
